+++ Tests/Unit/Validation/UserValidatorTest.php
@@ -6,6 +6,7 @@ namespace OliverKlee\Onetimeaccount\Tests\Unit\Validation;
 
 use OliverKlee\FeUserExtraFields\Domain\Model\FrontendUser;
 use OliverKlee\Onetimeaccount\Validation\UserValidator;
+use TYPO3\CMS\Core\SingletonInterface;
 use TYPO3\CMS\Extbase\Validation\Validator\AbstractValidator;
 use TYPO3\CMS\Extbase\Validation\Validator\ValidatorInterface;
 use TYPO3\TestingFramework\Core\Unit\UnitTestCase;
@@ -85,6 +86,14 @@ final class UserValidatorTest extends UnitTestCase
     {
         self::assertInstanceOf(ValidatorInterface::class, $this->subject);
         self::assertInstanceOf(AbstractValidator::class, $this->subject);
+    }
+
+    /**
+     * @test
+     */
+    public function isSingleton(): void
+    {
+        self::assertInstanceOf(SingletonInterface::class, $this->subject);
     }
 
     /**