+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Do not inject the `PasswordHashFactory` (#633)
 - Mark all injected classes as singletons (#631)
 
 ## 6.2.0